chore(ci): Report if table is in code but not in docs
