+++ scripts/missing_tables/go.mod
@@ -1,0 +1,5 @@
+module cloudquery/missingtables
+
+go 1.19
+
+require github.com/bmatcuk/doublestar/v4 v4.6.0
